+++ package.json
@@ -10,6 +10,7 @@
     "preview": "vite preview"
   },
   "dependencies": {
+    "axios": "^1.7.7",
     "eslint-config-prettier": "^9.1.0",
     "prop-types": "^15.8.1",
     "react": "^18.3.1",
